% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\name{counts_data_tutorial}
\alias{counts_data_tutorial}
\title{Gene Expression Counts Data}
\format{
A matrix with dimensions 19612 x 11.
}
\source{
Generated for tutorial purposes.
}
\description{
A matrix of gene expression counts with genes as rows and samples as columns.
}
