\name{cts_ERCC}
\docType{data}
\title{cts_ERCC Data}
\description{
  This dataset contains cts_ERCC spike-in controls for the HTGAnalyzer package.
}
\usage{cts_ERCC}
\format{
  A data frame with several variables:
    \describe{
      \item{\code{cts_ERCC_ID}}{Character: cts_ERCC identifier.}
      \item{\code{Concentration}}{Numeric: Concentration of the cts_ERCC spike-in controls.}
    }
}
\examples{
  data(cts_ERCC)
  head(cts_ERCC)
}
