% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{data_documentation}
\alias{data_documentation}
\title{Data Documentation for Tutorial Package}
\description{
This file contains documentation for the data sets included in the tutorial package.
The datasets provided are used for demonstrating various data analysis techniques.
}
\details{
The following datasets are included:
\itemize{
  \item \code{counts_data_tutorial}: A matrix of gene expression counts.
  \item \code{AnnotData_tutorial}: Annotation data for the samples.
  \item \code{outliers_tutorial}: A character vector of outlier sample IDs.
  \item \code{res_tutorial}: Results from differential expression analysis.
  \item \code{TME_data_tutorial}: Tumor microenvironment data including different cell types and scores.
}
}
\examples{
# Load data
data(counts_data_tutorial)
data(AnnotData_tutorial)
data(outliers_tutorial)
data(res_tutorial)
data(TME_data_tutorial)

# Inspect data
head(counts_data_tutorial)
head(AnnotData_tutorial)
head(outliers_tutorial)
head(res_tutorial)
head(TME_data_tutorial$EPIC)

}
\keyword{datasets}
