% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTG_analysis.R
\name{HTG_analysis}
\alias{HTG_analysis}
\title{HTG_analysis: Perform DESeq2 Analysis, GSEA, Convolution, and Survival Analysis}
\usage{
HTG_analysis(
  outliers,
  pattern = NULL,
  counts_data,
  col_data,
  design_formula = NULL,
  percentage_gene = 0.2,
  percentage_zero = 0.2,
  threshold_gene = 200,
  threshold_subject = 10,
  top_genes = c("CCND1", "MMP10", "CTTN"),
  heatmap_columns = NULL,
  contrast = NULL,
  pCutoff = 0.05,
  variable_01 = NULL,
  time = NULL,
  correlation = FALSE,
  dds = TRUE,
  generate_volcano = TRUE,
  remove_outliers = TRUE,
  GSEA = FALSE,
  generate_heatmap = TRUE,
  Convolution = TRUE,
  survival_analysis = FALSE,
  grupos = NULL
)
}
\arguments{
\item{outliers}{A character vector specifying the IDs of outlier samples to be removed. Outliers can also be identified using the HTG_QC function.}

\item{pattern}{(Optional) A regular expression pattern to identify control probes in the count data. For HTG, this could be "^NC-|^POS-|^GDNA-|^ERCC-". If NULL, the pattern will not be applied.}

\item{counts_data}{A matrix or data frame containing count data. It is recommended that the count data does not include control probes or outliers if not needed.}

\item{col_data}{A data frame containing sample annotations. For survival analysis, it must include variables `time` and `variable_01`.}

\item{design_formula}{The design formula for DESeq2 analysis, specified as a string without the tilde (~).}

\item{percentage_gene}{A numeric value between 0 and 1 indicating the minimum fraction of samples in which a gene must be expressed to be retained. Default is 0.2.}

\item{percentage_zero}{A numeric value between 0 and 1 indicating the maximum fraction of samples in which a gene can be zero to be retained. Default is 0.2.}

\item{threshold_gene}{Minimum count threshold per gene. Default is 200.}

\item{threshold_subject}{Minimum count threshold per subject. Default is 10.}

\item{top_genes}{A character vector specifying top genes for analysis. Default is c("CCND1", "MMP10", "CTTN").}

\item{heatmap_columns}{A character vector specifying the columns to be used for annotations in the heatmap. Default is c("design_formula", "Smoker").}

\item{contrast}{A character vector specifying the contrast for differential expression analysis. Default is c('column_name', 'variable1', 'variable2').}

\item{pCutoff}{The p-value cutoff for generating the volcano plot. Default is 0.05.}

\item{correlation}{A logical value indicating whether to perform correlation analysis within the convolution analysis. Default is FALSE.}

\item{dds}{A logical value indicating whether to perform DESeq2 analysis with filtering and without Lfc shrinkage. Default is TRUE.}

\item{generate_volcano}{A logical value indicating whether to generate a volcano plot. Default is TRUE.}

\item{remove_outliers}{A logical value indicating whether to remove outliers. Default is TRUE.}

\item{GSEA}{A logical value indicating whether to perform GSEA analysis. Default is FALSE.}

\item{generate_heatmap}{A logical value indicating whether to generate a heatmap. Default is TRUE.}

\item{Convolution}{A logical value indicating whether to perform convolution analysis. Default is TRUE.}

\item{survival_analysis}{A logical value indicating whether to perform survival analysis. Default is FALSE.}

\item{grupos}{A list specifying groups to be compared. Default is c("brain", "lymph node").}
}
\value{
Returns an object with the results of the specified contrast and saves an Excel file with the results, along with PDF files of the generated plots.
}
\description{
This function conducts a comprehensive analysis pipeline including DESeq2 differential expression analysis (DEA), Gene Set Enrichment Analysis (GSEA), convolution analysis, and survival analysis. The pipeline supports optional steps for generating volcano plots and heatmaps. The function is suitable for both HTG and RNA-seq data.
}
\examples{
ALL_done<- HTG_analysis(outliers = outliers, pattern = "^NC-|^POS-|^GDNA-|^ERCC-", counts_data, col_data = AnnotData,
design_formula = "Ciclina2" , percentage_gene = 0.2, percentage_zero = 0.2,
threshold_gene = 200, threshold_subject = 10, top_genes = c("CCND1", "MMP10", "CTTN"), heatmap_columns = c("Ciclina2", "Smoker"),
contrast = c("Ciclina2", "high", "low"), pCutoff = 5e-2,variable_01 = "smoke_01", time = "time",
correlation = TRUE,
dds = TRUE, generate_volcano = TRUE, remove_outliers = TRUE, GSEA = TRUE, generate_heatmap = TRUE, Convolution = TRUE,
survival_analysis = TRUE, grupos = c("high", "low"))


}
