% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.4- calculate_ratios.R
\name{HTG_calculate_ratios}
\alias{HTG_calculate_ratios}
\title{HTG_calculate_ratios}
\usage{
HTG_calculate_ratios(
  counts_filtered,
  cts_POS,
  cts_NC,
  cts_GDNA,
  cts_ERCC,
  save_csv = FALSE,
  csv_file = "ratios.csv"
)
}
\arguments{
\item{counts_filtered}{A data frame containing filtered counts data (counts data excluding probes).}

\item{cts_POS}{A data frame containing counts data for positive controls.}

\item{cts_NC}{A data frame containing counts data for negative controls.}

\item{cts_GDNA}{A data frame containing counts data for genomic DNA.}

\item{cts_ERCC}{A data frame containing counts data for ERCC controls.}

\item{save_csv}{Logical, indicating whether to save the calculated ratios as a CSV file. Default is FALSE.}

\item{csv_file}{Character string specifying the file name if `save_csv` is TRUE. Default is "ratios.csv".}
}
\value{
A data frame containing:
- Total counts: Sum of counts for positive controls, genomic DNA, negative controls, and ERCC controls.
- Ratios: Ratios of positive controls, negative controls, and ERCC controls relative to total counts.
- Summary statistics:For each sample: median, minimum, maximum, standard deviation (SD), first quartile (Q1), and third quartile (Q3).
}
\description{
This function calculates various ratios based on counts data for different categories such as positive controls, genomic DNA, negative controls, and ERCC controls. Additionally, it computes summary statistics including median, minimum, maximum, standard deviation, first quartile (Q1), and third quartile (Q3) for each sample.
}
\examples{
HTG_calculate_ratios(counts_filtered, cts_POS, cts_NC, cts_GDNA, cts_ERCC,
                     save_csv = TRUE, csv_file = "ratios.csv")
}
