% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.2- HTG_calculate_summary_stats.R
\name{HTG_calculate_summary_stats}
\alias{HTG_calculate_summary_stats}
\title{HTG_calculate_summary_stats}
\usage{
HTG_calculate_summary_stats(counts_filtered)
}
\arguments{
\item{counts_filtered}{A data frame containing the input counts.}
}
\value{
A data frame containing summary statistics for each column of the input data.
}
\description{
This function calculates summary statistics for each column of the input data frame, including minimum, maximum, mean, median, mode, standard deviation, variance, range, first quartile (Q1), third quartile (Q3), interquartile range (IQR), skewness, kurtosis, count of missing values, and coefficient of variation (CV).
}
\examples{
summary<- HTG_calculate_summary_stats(counts_filtered)
}
