% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-process_DESeq.R
\name{HTG_DESeq}
\alias{HTG_DESeq}
\title{Process DESeq2 analysis}
\usage{
HTG_DESeq(
  count_data,
  col_data,
  design_formula,
  threshold_gene = 200,
  threshold_subject = 10,
  heatmap_columns = c("GROUP", "RUN"),
  contrast = c("column_name", "variable1", "variable2"),
  pCutoff = 0.05
)
}
\arguments{
\item{count_data}{A matrix or data frame containing count data. It is recommended that the count data frame does not include probes nor outliers.}

\item{col_data}{A data frame containing sample annotations.}

\item{design_formula}{The design formula for DESeq2 analysis without the tilde (~).}

\item{threshold_gene}{Minimum count threshold per gene.}

\item{threshold_subject}{Minimum count threshold per subject.}

\item{heatmap_columns}{A character vector specifying the columns to be used for annotations in the heatmap.}

\item{contrast}{A character vector specifying the contrast for differential expression analysis.}

\item{pCutoff}{The p-value cutoff for generating the volcano plot.}
}
\value{
Returns an object with the results of contrast demanded and an excel with the results.
}
\description{
This function processes DESeq2 analysis with customizable parameters.
}
\examples{
dds_processed <- process_DESeq(counts_filtered, AnnotData, "column_name", threshold_gene = 200, threshold_subject = 10, heatmap_columns = c("GROUP", "RUN"), contrast = c('column_name', 'variable1','variable2'), pCutoff = 5e-2)
}
