% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.1- HTG_filterCounts.R
\name{HTG_filter}
\alias{HTG_filter}
\title{HTG_filter}
\usage{
HTG_filter(data_frame, pattern = "^NC-|^POS-|^GDNA-|^ERCC-")
}
\arguments{
\item{data_frame}{A data frame containing the data to be filtered.}

\item{pattern}{A string specifying the pattern of prefixes to filter out. The default pattern is "^NC-|^POS-|^GDNA-|^ERCC-", which will remove rows starting with these prefixes.}
}
\value{
A filtered data frame with rows that match the specified prefixes removed.
}
\description{
This function filters out rows in a data frame that have specified prefixes in their row names.
}
\examples{
# Filter out rows with default prefixes
counts_filtered <- HTG_filter(counts_data)

# Filter out rows with specific prefixes
counts_filtered <- HTG_filter(counts_data, "^NC-|^POS-")

# Filter out rows with a specific pattern
counts_filtered <- HTG_filter(counts_data, "A1BG")
}
