% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-2-filterCounts.R
\name{HTG_filterCounts}
\alias{HTG_filterCounts}
\title{Filter Counts Data}
\usage{
HTG_filterCounts(counts)
}
\arguments{
\item{counts}{A data frame containing counts data with genes as rows and samples as columns.}
}
\value{
A filtered counts data frame with specified prefixes removed.
}
\description{
This function filters counts data to remove rows with specific prefixes such as "NC-", "POS-", "GDNA-", and "ERCC-".
}
\examples{
counts_filtered <- HTG_filterCounts(counts)
}
