% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05- HTG_GSEA.R
\name{HTG_GSEA}
\alias{HTG_GSEA}
\title{HTG_GSEA}
\usage{
HTG_GSEA(res)
}
\arguments{
\item{res}{The DESeq2 results object containing log2 fold changes and statistical information.}

\item{gene_list}{A vector of log2 fold changes representing gene expression changes used for GSEA.}

\item{kegg_gene_list}{A vector of log2 fold changes representing KEGG gene expression changes used for KEGG analysis.}
}
\value{
This function does not explicitly return a value. It saves plots and tables related to GSEA results in the "GSE_results" folder and generates two Excel files containing the results of `gene_list` and `kegg_gene_list`.
}
\description{
This function conducts GSEA to identify enriched gene sets based on DESeq2 log2 fold changes.
}
\details{
Perform Gene Set Enrichment Analysis (GSEA) using DESeq2 results, generating various plots and saving results.


This function performs GSEA using DESeq2 results (`res`) to identify enriched gene sets. It generates several plots including Dotplot, Emaplot, and Ridgeplot for both general enrichment and KEGG pathways. The plots are saved in a PDF file named "GSEA_analysis_plots.pdf" in the working directory. If `res` does not contain significant results (padj < 0.05), the function will not generate certain plots and will display an error message.
}
\examples{
HTG_GSEA(res)
}
