% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-GSEA.R
\name{HTG_GSEAresults}
\alias{HTG_GSEAresults}
\title{Process GSEA Results}
\usage{
HTG_GSEAresults(res)
}
\arguments{
\item{res}{The DESeq2 results object containing log2 fold changes.}

\item{gene_list}{A vector of log2 fold changes representing gene expression changes.}

\item{kegg_gene_list}{A vector of log2 fold changes representing KEGG gene expression changes.}
}
\value{
This function doesn't return any value explicitly. It saves plots and tables related to GSEA results in the "GSE_results" folder. Additionally, it generates two Excel files containing the results of gene_list and KEGG gene list.
}
\description{
This function processes GSE (Gene Set Enrichment) results obtained from DESeq2 analysis.
}
\examples{
HTG_GSEA_results(res)
}
