% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateHeatmap.R
\name{HTG_HeatmapQC}
\alias{HTG_HeatmapQC}
\title{HTG_HeatmapQC}
\usage{
HTG_HeatmapQC(
  ratios,
  counts_filtered,
  n_samples = 3,
  threshold_pos = 4,
  threshold_gdna = 0.02,
  threshold_neg = 0.05,
  threshold_size = 7e+06,
  threshold_median = 5,
  threshold_pca = 1,
  threshold_ERCC = 0.05
)
}
\arguments{
\item{ratios}{A data frame containing QC ratios and other relevant metrics.}

\item{counts_filtered}{A matrix of counts data with samples as columns and features as rows.}

\item{n_samples}{Number of samples to label as outliers in PCA.}

\item{threshold_pos}{Threshold for positive QC value.}

\item{threshold_gdna}{Threshold for gdna QC value.}

\item{threshold_neg}{Threshold for negative QC value.}

\item{threshold_size}{Threshold for library size.}

\item{threshold_median}{Threshold for median.}

\item{threshold_pca}{Threshold for PCA genes.}

\item{threshold_ERCC}{Threshold for ERCC QC value.}
}
\value{
A heatmap plot displaying QC ratios.
}
\description{
Generate a heatmap based on QC ratios
}
\examples{
HTG_HeatmapQC(ratios, counts_filtered, n_samples = 3)
}
