% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-inHTG.R
\name{HTG_import}
\alias{HTG_import}
\title{Import HTG counts data from an Excel file}
\usage{
HTG_import(file_path)
}
\arguments{
\item{file_path}{Path to the Excel file containing HTG counts}
}
\value{
A data frame with HTG counts data
}
\description{
Import HTG counts data from an Excel file
}
\examples{
htg_data <- inHTG("path/to/your/excel/file.xlsx")

}
