% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01- HTG_import_counts.R
\name{HTG_import_counts}
\alias{HTG_import_counts}
\title{HTG_import_counts}
\usage{
HTG_import_counts(file_path, file_type)
}
\arguments{
\item{file_path}{Path to the Excel file containing HTG counts. All the columns have to had a name in excel file.}

\item{file_type}{Type of file being imported, either "HTG" or "RNAseq".}
}
\value{
A data frame with HTG counts data.
}
\description{
Import counts data from an Excel file. It can be either HTG excel or RNAseq.
}
\examples{
\dontrun{
# Replace "path/to/your/excel/file.xlsx" with the actual path to your Excel file
path <- "path/to/your/excel/file.xlsx"
htg_data <- HTG_import_counts(path, "HTG")
rna_data <- HTG_import_counts(path, "RNAseq")
}
}
