% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03.1- plotPCA_genes.R
\name{HTG_plotPCA_genes}
\alias{HTG_plotPCA_genes}
\title{HTG_plotPCA_genes}
\usage{
HTG_plotPCA_genes(counts_filtered, n_samples = 3)
}
\arguments{
\item{counts_filtered}{A data frame containing filtered counts data with genes as rows and samples as columns.}

\item{n_samples}{The number of samples to label as outliers based on their distance from the center of the PCA plot.}
}
\value{
A combined plot consisting of the PCA plot, explained variability plot, and accumulated explained variability plot.
}
\description{
This function generates a PCA plot specifically for genes, along with plots showing explained variability and accumulated explained variability. It highlights the samples that are the most distant from the center of the PCA plot.
}
\examples{
HTG_plotPCA_genes(counts_filtered, n_samples = 3)
}
