% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6-plotPCA_probes.R
\name{HTG_plotPCA_probes}
\alias{HTG_plotPCA_probes}
\title{PCA Plot and Explained Variability}
\usage{
HTG_plotPCA_probes(counts, n_samples = 3)
}
\arguments{
\item{counts}{An object with sample names as column names and genes as row names, containing counts data.}

\item{n_samples}{The number of samples to label in red, which are farthest from the center.}
}
\value{
A combined plot of PCA samples, the first 10 principal components, and accumulated explained variability.
}
\description{
This function generates a PCA plot of samples and additional plots showing the explained variability of the principal components.
}
\examples{
plotPCA_probes(counts_data, 5)
plotPCA_probes(counts_data)
}
