% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTG_QC.R
\name{HTG_QC}
\alias{HTG_QC}
\title{HTG_QC: Quality Control for HTG EdgeSeq Data}
\arguments{
\item{counts_data}{A data frame containing the HTG count data. The data must include probes that start with "^NC-|^POS-|^GDNA-|^ERCC-" for the function to work correctly.}

\item{pattern}{A regular expression pattern to identify control probes in the count data. For HTG data, this could be "^NC-|^POS-|^GDNA-|^ERCC-". If NULL, the pattern will not be applied.}

\item{threshold_superior_pos}{Threshold for upper limit of positive control ratio.}

\item{threshold_inferior_pos}{Threshold for lower limit of positive control ratio.}

\item{threshold_line_pos}{Threshold line for positive control ratio.}

\item{threshold_superior_lib}{Threshold for upper limit of library size.}

\item{threshold_inferior_lib}{Threshold for lower limit of library size.}

\item{threshold_lib}{Threshold line for library size.}

\item{threshold_superior_nc}{Threshold for upper limit of negative control ratio.}

\item{threshold_inferior_nc}{Threshold for lower limit of negative control ratio.}

\item{threshold_line_nc}{Threshold line for negative control ratio.}

\item{threshold_superior_gdna}{Threshold for upper limit of genomic DNA ratio.}

\item{threshold_inferior_gdna}{Threshold for lower limit of genomic DNA ratio.}

\item{threshold_line_gdna}{Threshold line for genomic DNA ratio.}

\item{threshold_superior_ercc}{Threshold for upper limit of ERCC control ratio.}

\item{threshold_inferior_ercc}{Threshold for lower limit of ERCC control ratio.}

\item{threshold_line_ercc}{Threshold line for ERCC control ratio.}

\item{threshold_superior_median}{Threshold for upper limit of median ratio.}

\item{threshold_inferior_median}{Threshold for lower limit of median ratio.}

\item{threshold_line_median}{Threshold line for median ratio.}

\item{n_samples}{Number of samples to label as outliers in plots.}

\item{save_csv}{Logical, whether to save the ratios as a CSV file. Default is FALSE.}

\item{csv_file}{The name of the CSV file to save the ratios if save_csv is TRUE. Default is "QC_results.csv".}
}
\value{
This function generates multiple plots displaying various QC metrics, including PCA and violin plots, and saves an Excel file with all the ratios. Additionally, it identifies and returns the most probable outliers based on the QC analysis.
}
\description{
This function performs various quality control (QC) checks for HTG EdgeSeq transcriptomic panel data. The QC checks include:

QC0: Percentage of positive values < than 4%; QC1: Library size > than 7e+06; QC2: Negative control threshold < than 0.045; QC3: Genomic DNA threshold < than 0.02; QC4: ERCC threshold < than 0.025; Median:threshold > 5.

In addition to these plots, highlighting potential outlier samples. The function also creates a data frame including the sum of each probe for each sample (total genes, positive, negative, gdna, and ercc)
the ratio for each sample and the size of each sample. Additionally, a statistical .csv is generated with columns for Min, Max, Mean, Median, Mode, SD, Variance, Range, Q1, Q3, IQR, Skewness, Kurtosis, Missing, and CV.

This function also includes an optional heatmap to highlight potential outlier samples, which will be saved in a vector.

The plots will saved in the current working directory.
}
\examples{
# Run the function with example data
HTG_QC(counts_data = counts_data_tutorial, pattern = "^NC-|^POS-|^GDNA-|^ERCC-", save_csv = TRUE)

}
