% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8-remove_outliers_COUNTS.R
\name{HTG_remove_outliers_COUNTS}
\alias{HTG_remove_outliers_COUNTS}
\title{Remove outliers from counts data based on specified sample IDs}
\usage{
HTG_remove_outliers_COUNTS(counts_data, outlier_rows)
}
\arguments{
\item{counts_data}{A data frame containing counts data.}

\item{outlier_rows}{A vector of sample IDs (column names) to be removed from the counts data frame.}
}
\value{
Returns a data frame with the outlier columns removed.
}
\description{
This function removes specified sample IDs from a data frame containing counts data.
}
\examples{
outlier_rows <- c("B04-23456-B", "B16-39493-1")
cleaned_counts <- remove_outliers_COUNTS(counts_data, outlier_rows)
}
