% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.3- HTG_subset.R
\name{HTG_subset}
\alias{HTG_subset}
\title{HTG_subset}
\usage{
HTG_subset(data_frame, prefix)
}
\arguments{
\item{data_frame}{A data frame from which rows need to be subsetted.}

\item{prefix}{A string specifying the prefix used to match and subset rows from the data frame.}
}
\value{
A subset of the data frame containing rows with row names that match the specified prefix.
}
\description{
This function subsets a data frame based on a specified prefix in the row names. It allows you to extract specific rows that match a given pattern.
}
\examples{
# Subset rows with the prefix "ERCC"
cts_ERCC <- HTG_subset(counts_data, "ERCC")

# Subset rows with the prefix "POS"
cts_POS <- HTG_subset(counts_data, "POS")

# Subset rows with the prefix "NC"
cts_NC <- HTG_subset(counts_data, "NC")

# Subset rows with the prefix "GDNA"
cts_GDNA <- HTG_subset(counts_data, "GDNA")
}
