% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-1-subset_counts.R
\name{HTG_subset_counts}
\alias{HTG_subset_counts}
\title{Subset Counts Data Frame}
\usage{
HTG_subset_counts(counts, prefix)
}
\arguments{
\item{counts}{A data frame containing counts data.}

\item{prefix}{A prefix string used to subset rows from the counts data frame.}
}
\value{
A subset of the counts data frame containing rows with row names that match the specified prefix.
}
\description{
This function subsets a counts data frame based on a specified prefix in row names.
}
\examples{
cts_ERCC <- subset_counts(counts, "ERCC")
}
