% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07- survival_function.R
\name{HTG_survival}
\alias{HTG_survival}
\title{Perform Survival Analysis}
\usage{
HTG_survival(
  variable_01,
  time,
  col_data,
  count_data,
  res = NULL,
  genes_to_use = NULL,
  outliers,
  pattern = NULL,
  remove_outliers = TRUE,
  TME = NULL
)
}
\arguments{
\item{variable_01}{Character. The name of the survival event variable.}

\item{time}{Character. The name of the time variable.}

\item{res}{Data frame. Result data with padj values.}

\item{outliers}{outliers that you want to be deleted.}

\item{pattern}{patter to delete. "^NC-|^POS-|^GDNA-|^ERCC-"}

\item{remove_outliers}{If you have the outliers and want to delete. It's set TRUE.}

\item{survival_analysis}{Logical. If TRUE, performs survival analysis.}

\item{AnnotData}{Data frame. Annotation data.}

\item{counts_filtered}{Matrix. Filtered count data.}
}
\value{
Generates survival analysis plots and saves them as PDF files.
}
\description{
This function performs survival analysis based on the provided data.
}
\examples{
HTG_survival(variable_01 = "smoke_01",time = "time", col_data = AnnotData, genes_to_use = NULL,
count_data = counts_data , res = res, outliers= outliers, pattern = "^NC-|^POS-|^GDNA-|^ERCC-", remove_outliers = TRUE )

}
