% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTG_TME.R
\name{HTG_TME}
\alias{HTG_TME}
\title{HTG_TME}
\arguments{
\item{outliers}{Vector of outlier sample IDs to be removed from analysis.}

\item{pattern}{Regular expression pattern to identify outlier sample IDs. For HTG, this could be "^NC-|^POS-|^GDNA-|^ERCC-". If NULL, the pattern will not be applied.}

\item{counts_data}{Count data matrix of gene expression.}

\item{AnnotData}{DataFrame containing sample metadata.}

\item{design_formula}{Formula specifying the design of the analysis.}

\item{DEA}{Preprocessed DESeqDataSet for differential expression analysis.}

\item{generate_volcano}{Logical; whether to generate volcano plots.}

\item{remove_outliers}{Logical; whether to remove outlier samples.}
}
\value{
The function outputs multiple CSV files (TPM normalized counts and Results of TME methods EPIC, quanTIseq, and xCell), PDF with graphics and return a dataframe with the three TME methods.
}
\description{
This function carries out a comprehensive TME analysis involving TPM normalization and TME using multiple methods (EPIC, quanTIseq, and xCell). It produces several output files containing the results of the normalization and TME processes.
}
\examples{
TME_results <- HTG_TME(
  outliers = outliers_tutorial,
  pattern = "^NC-|^POS-|^GDNA-|^ERCC-",
  counts_data = counts_data_tutorial,
  AnnotData = AnnotData_tutorial,
  design_formula = "HPV_status",
  remove_outliers = TRUE,
  DEA = NULL
)

}
