% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-package.R
\name{HTGAnalyzer-package}
\alias{HTGAnalyzer-package}
\title{HTGAnalyzer Package}
\description{
This package performs HTG analysis and various associated tasks.
}
\details{
This file centralizes the import statements for the HTGAnalyzer package.
It ensures that functions from the imported packages are available throughout the package.
}
