% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\name{outliers_tutorial}
\alias{outliers_tutorial}
\title{Outlier Sample IDs}
\format{
A character vector with sample IDs.
}
\source{
Generated for tutorial purposes.
}
\description{
A character vector containing the IDs of samples identified as outliers.
}
