% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\name{TME_data_tutorial}
\alias{TME_data_tutorial}
\title{Tumor Microenvironment Data}
\format{
A list with the following components:
\itemize{
  \item \code{EPIC}: Data frame with cell type abundances estimated by EPIC.
  \item \code{QTI}: Data frame with cell type abundances estimated by QTI.
  \item \code{XCELL}: Data frame with cell type abundances estimated by XCELL.
}
}
\source{
Generated for tutorial purposes.
}
\description{
A list containing different tumor microenvironment scores and cell type abundances.
}
