% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTG_analysis.R
\name{HTG_analysis}
\alias{HTG_analysis}
\title{HTG_analysis: Perform DESeq2 Analysis, GSEA, TME, and Survival Analysis}
\arguments{
\item{outliers}{A character vector specifying the IDs of outlier samples to be removed. If you have RNA-seq data or don't have outliers, this parameter can be set to NULL. Note: For RNA-seq data, `remove_outliers` should be set to FALSE.}

\item{pattern}{A regular expression pattern to identify control probes in the count data. For HTG, this could be "^NC-|^POS-|^GDNA-|^ERCC-". If NULL, the pattern will not be applied.}

\item{counts_data}{A matrix or data frame containing count data. It is recommended that the count data does not include control probes or outliers if not needed.}

\item{col_data}{A data frame containing sample annotations. For survival analysis, it must include variables `time` and `variable_01`.}

\item{design_formula}{The design formula for DESeq2 analysis, specified as a string without the tilde (~).}

\item{threshold_gene}{Minimum count threshold per gene. Default is 200.}

\item{threshold_subject}{Minimum count threshold per subject. Default is 10.}

\item{heatmap_columns}{A character vector specifying the columns to be used for annotations in the heatmap. It is recommended to include the column used in `design_formula`. Default is NULL.}

\item{contrast}{A character vector specifying the contrast for differential expression analysis. Should include the column name and two levels for comparison.}

\item{pCutoff}{The p-value cutoff for generating the volcano plot. Default is 0.05.}

\item{remove_outliers}{A logical value indicating whether to remove outliers. Default is TRUE. Note: For RNA-seq data, this should be set to FALSE.}

\item{GSEA}{A logical value indicating whether to perform Gene Set Enrichment Analysis. Default is FALSE.}

\item{generate_heatmap}{A logical value indicating whether to generate a heatmap. Default is TRUE.}

\item{TME}{A logical value indicating whether to perform Tumor Microenvironment (TME) analysis. Default is TRUE.}

\item{survival_analysis}{A logical value indicating whether to perform survival analysis. Default is FALSE.}

\item{percentage_gene}{A numeric value between 0 and 1 indicating the minimum fraction of samples in which a gene must be expressed to be retained. Default is 0.2.}

\item{percentage_zero}{A numeric value between 0 and 1 indicating the maximum fraction of samples in which a gene can be zero to be retained. Default is 0.2.}

\item{genes_to_use}{A character vector specifying top genes for analysis. Default is c("CCND1", "MMP10", "CTTN").}

\item{DEA}{A logical value indicating whether to perform DESeq2 analysis with filtering and without LFC shrinkage. Default is TRUE.}

\item{variable_01}{A character string specifying the event/censoring variable used in the survival analysis.}

\item{time}{A character string specifying the time variable used in the survival analysis.}
}
\value{
Returns an object with the results of the specified contrast and saves an Excel file with the results, along with PDF files of the generated plots.
}
\description{
This function conducts a comprehensive analysis pipeline including DESeq2 differential expression analysis (DEA), Gene Set Enrichment Analysis (GSEA), TME analysis, and survival analysis. The pipeline supports optional steps for generating volcano plots and heatmaps. The function is suitable for both HTG and RNA-seq data.
}
\examples{
ALL_done <- HTG_analysis(
  outliers = outliers_tutorial,
  pattern = "^NC-|^POS-|^GDNA-|^ERCC-",
  counts_data = counts_data_tutorial,
  col_data = AnnotData_tutorial,
  design_formula = "HPV_status",
  percentage_gene = 0.2,
  percentage_zero = 0.2,
  threshold_gene = 200,
  threshold_subject = 10,
  genes_to_use = c("CCND1", "MMP10", "CTTN"),
  heatmap_columns = c("HPV_status", "Ciclina_D1"),
  contrast = c("HPV_status", "Positive", "Negative"),
  pCutoff = 5e-2,
  variable_01 = "Recurrence_01",
  time = "Time_to_death_surv",
  DEA = TRUE,
  remove_outliers = TRUE,
  GSEA = TRUE,
  generate_heatmap = TRUE,
  TME = TRUE,
  survival_analysis = TRUE
)
}
