% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-calculate_ratios.R, R/3plotControls.R
\name{HTG_calculate_ratios}
\alias{HTG_calculate_ratios}
\alias{HTG_plotControls}
\title{Calculate Ratios}
\usage{
HTG_calculate_ratios(
  counts_filtered,
  cts_POS,
  cts_NC,
  cts_GDNA,
  cts_ERCC,
  save_csv = FALSE,
  csv_file = "ratios.csv"
)

HTG_plotControls(
  counts_filtered,
  ratios,
  threshold_superior_pos = 5,
  threshold_inferior_pos = 3,
  threshold_line_pos = 4,
  threshold_superior_lib = 8e+06,
  threshold_inferior_lib = 5e+06,
  threshold_lib = 7e+06,
  threshold_superior_nc = 0.05,
  threshold_inferior_nc = 0.035,
  threshold_line_nc = 0.045,
  threshold_superior_gdna = 0.025,
  threshold_inferior_gdna = 0.015,
  threshold_line_gdna = 0.02,
  threshold_superior_ercc = 0.03,
  threshold_inferior_ercc = 0.015,
  threshold_line_ercc = 0.025,
  threshold_superior_median = 7,
  threshold_inferior_median = 3,
  threshold_line_median = 5
)
}
\arguments{
\item{counts_filtered}{data frame without probes.}

\item{cts_POS}{A data frame containing counts data for positive controls.}

\item{cts_NC}{A data frame containing counts data for negative controls.}

\item{cts_GDNA}{A data frame containing counts data for genomic DNA.}

\item{cts_ERCC}{A data frame containing counts data for ERCC controls.}

\item{save_csv}{Logical, indicating whether to save the calculated ratios as a CSV file. Default is FALSE.}

\item{csv_file}{Character string specifying the file name if `save_csv` is TRUE. Default is "ratios.csv".}

\item{ratios}{A data frame containing QC ratios and other relevant metrics.}

\item{threshold_superior_pos}{Threshold for positive control upper limit.}

\item{threshold_inferior_pos}{Threshold for positive control lower limit.}

\item{threshold_line_pos}{Threshold line for positive control.}

\item{threshold_superior_nc}{Threshold for negative control upper limit.}

\item{threshold_inferior_nc}{Threshold for negative control lower limit.}

\item{threshold_line_nc}{Threshold line for negative control.}

\item{threshold_superior_gdna}{Threshold for genomic DNA upper limit.}

\item{threshold_inferior_gdna}{Threshold for genomic DNA lower limit.}

\item{threshold_line_gdna}{Threshold line for genomic DNA.}

\item{threshold_superior_ercc}{Threshold for ERCC control upper limit.}

\item{threshold_inferior_ercc}{Threshold for ERCC control lower limit.}

\item{threshold_line_ercc}{Threshold line for ERCC control.}

\item{threshold_superior_median}{Threshold for median upper limit.}

\item{threshold_inferior_median}{Threshold for median lower limit.}

\item{threshold_line_median}{Threshold line for median.}
}
\value{
A data frame containing calculated ratios for different categories.

Multiple plots displaying different QC metrics.
}
\description{
This function calculates ratios based on counts data for different categories such as positive controls, genomic DNA, etc.

This function generates multiple plots to visualize various quality control (QC) metrics.
}
\examples{
calculate_ratios(counts_filtered, cts_POS, cts_NC, cts_GDNA, cts_ERCC, save_csv = TRUE, csv_file = "ratios.csv")
plotControls(counts_filtered,ratios)
}
