% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7-calculate_summary_stats.R
\name{HTG_calculate_summary_stats}
\alias{HTG_calculate_summary_stats}
\title{Calculate Summary Statistics}
\usage{
HTG_calculate_summary_stats(data)
}
\arguments{
\item{data}{A data frame containing the input data.}
}
\value{
A data frame containing summary statistics for each column of the input data.
}
\description{
This function calculates summary statistics including minimum, maximum, mean, and median for each column of the input data.
}
\examples{
calculate_summary_stats(data)
}
