% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTG_calculate_summary_stats.R
\name{HTG_calculate_summary_stats}
\alias{HTG_calculate_summary_stats}
\title{HTG_calculate_summary_stats}
\usage{
HTG_calculate_summary_stats(counts_data, pattern = NULL)
}
\arguments{
\item{counts_data}{A data frame containing the input counts.}

\item{pattern}{(Optional) A regular expression pattern to identify control probes in the count data. For HTG, this could be "^NC-|^POS-|^GDNA-|^ERCC-". If NULL, the pattern will not be applied.}
}
\value{
A data frame containing summary statistics for each column of the input data. and also will be .csv
}
\description{
This function calculates summary statistics for each column of the input data frame, including minimum, maximum, mean, median, mode, standard deviation, variance, range, first quartile (Q1), third quartile (Q3), interquartile range (IQR), skewness, kurtosis, count of missing values, and coefficient of variation (CV).
}
\examples{
summary <- HTG_calculate_summary_stats(counts_data_tutorial, pattern = "^NC-|^POS-|^GDNA-|^ERCC-")
}
