% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.2- HTG_calculate_summary_stats.R
\name{HTG_calculate_summary_stats}
\alias{HTG_calculate_summary_stats}
\title{HTG_calculate_summary_stats}
\usage{
HTG_calculate_summary_stats(counts_filtered)
}
\arguments{
\item{counts_filtered}{A counts_filtered frame containing the input counts_filtered.}
}
\value{
A counts_filtered frame containing summary statistics for each column of the input counts_filtered.
}
\description{
This function calculates summary statistics including minimum, maximum, mean, and median for each column of the input data.
}
\examples{
summary<- HTG_calculate_summary_stats(counts_filtered)
}
