% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTG_DEA.R
\name{HTG_DEA}
\alias{HTG_DEA}
\title{HTG_DEA}
\arguments{
\item{outliers}{A character vector of sample IDs to be removed as outliers.}

\item{counts_data}{A matrix or data frame of raw count data, with genes as rows and samples as columns. Probes will be deleted.}

\item{col_data}{A data frame containing the sample metadata. Must include an 'id' column that matches the column names of `counts_data`.}

\item{design_formula}{A character string representing the design formula for DESeq2. Must specify two groups for comparison.}

\item{heatmap_columns}{A character vector of columns in `col_data` to be used for heatmap annotation. It is recommended to include the column used in `design_formula`.}

\item{contrast}{A character vector specifying the contrast to be used for differential expression analysis. Should include two levels for comparison.}

\item{pattern}{A regular expression pattern to filter out unwanted genes from the count data. For HTG, this could be "^NC-|^POS-|^GDNA-|^ERCC-". If NULL, the pattern will not be applied.}

\item{remove_outliers}{A logical value indicating whether to remove outliers specified in `outliers`.}

\item{percentage_gene}{A numeric value specifying the minimum percentage of samples a gene must be expressed in to be kept.}

\item{percentage_zero}{A numeric value specifying the maximum percentage of samples a gene can be zero in to be kept.}

\item{threshold_gene}{A numeric value specifying the minimum count for a gene to be considered expressed.}

\item{threshold_subject}{A numeric value specifying the minimum number of subjects a gene must be expressed in to be kept.}

\item{pCutoff}{A numeric value specifying the p-value cutoff for significance in the volcano plot.}

\item{apply_filtering}{A logical value indicating whether to apply filtering of genes based on expression thresholds.}

\item{apply_lfc_shrinkage}{A logical value indicating whether to apply log fold change shrinkage.}

\item{extract_shrinkage}{A logical value indicating whether to extract shrinkage results.}
}
\value{
A DESeqResults object containing the differential expression results. If `apply_lfc_shrinkage` is TRUE, the function returns the results with log fold change shrinkage applied.
}
\description{
Perform differential expression analysis using DESeq2, with options for filtering and log fold change (LFC) shrinkage.
}
\examples{
results <- HTG_DEA(
  outliers = outliers_tutorial,
  counts_data = counts_data_tutorial,
  col_data = AnnotData_tutorial,
  design_formula = "HPV_status",
  heatmap_columns = c("HPV_status", "Recurrence"),
  contrast = c("HPV_status", "Positive", "Negative"),
  pattern = "^NC-|^POS-|^GDNA-|^ERCC-",
  remove_outliers = TRUE,
  percentage_gene = 0.2,
  percentage_zero = 0.2,
  threshold_gene = 200,
  threshold_subject = 10,
  pCutoff = 5e-2,
  apply_filtering = TRUE,
  apply_lfc_shrinkage = TRUE,
  extract_shrinkage = FALSE
)
}
