% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCA_genes.R
\name{HTG_plotPCA_genes}
\alias{HTG_plotPCA_genes}
\title{Plot PCA for Genes Only}
\usage{
HTG_plotPCA_genes(counts_filtered, n_samples = 3)
}
\arguments{
\item{counts_filtered}{A data frame containing filtered counts data with genes as rows and samples as columns.}

\item{n_samples}{Number of samples to label as outliers.}
}
\value{
A combined plot consisting of PCA, explained variability, and accumulated explained variability plots.
}
\description{
This function generates a PCA plot for genes only, along with the explained variability and accumulated explained variability plots.
}
\examples{
HTG_plotPCA_genes(counts_filtered, n_samples = 3)
}
