% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.0- HTG_QC.R
\name{HTG_QC}
\alias{HTG_QC}
\title{HTG_QC}
\usage{
HTG_QC(
  counts_data,
  threshold_superior_pos = 5,
  threshold_inferior_pos = 3,
  threshold_line_pos = 4,
  threshold_superior_lib = 8e+06,
  threshold_inferior_lib = 5e+06,
  threshold_lib = 7e+06,
  threshold_superior_nc = 0.05,
  threshold_inferior_nc = 0.035,
  threshold_line_nc = 0.045,
  threshold_superior_gdna = 0.025,
  threshold_inferior_gdna = 0.015,
  threshold_line_gdna = 0.02,
  threshold_superior_ercc = 0.03,
  threshold_inferior_ercc = 0.015,
  threshold_line_ercc = 0.025,
  threshold_superior_median = 7,
  threshold_inferior_median = 3,
  threshold_line_median = 5,
  n_samples = 3,
  save_csv = FALSE,
  csv_file = "QC_results.csv",
  show_heatmap = FALSE
)
}
\arguments{
\item{threshold_superior_pos}{Threshold for upper limit of positive control ratio.}

\item{threshold_inferior_pos}{Threshold for lower limit of positive control ratio.}

\item{threshold_line_pos}{Threshold line for positive control ratio.}

\item{threshold_superior_lib}{Threshold for upper limit of library size.}

\item{threshold_inferior_lib}{Threshold for lower limit of library size.}

\item{threshold_lib}{Threshold line for library size.}

\item{threshold_superior_nc}{Threshold for upper limit of negative control ratio.}

\item{threshold_inferior_nc}{Threshold for lower limit of negative control ratio.}

\item{threshold_line_nc}{Threshold line for negative control ratio.}

\item{threshold_superior_gdna}{Threshold for upper limit of genomic DNA ratio.}

\item{threshold_inferior_gdna}{Threshold for lower limit of genomic DNA ratio.}

\item{threshold_line_gdna}{Threshold line for genomic DNA ratio.}

\item{threshold_superior_ercc}{Threshold for upper limit of ERCC control ratio.}

\item{threshold_inferior_ercc}{Threshold for lower limit of ERCC control ratio.}

\item{threshold_line_ercc}{Threshold line for ERCC control ratio.}

\item{threshold_superior_median}{Threshold for upper limit of median ratio.}

\item{threshold_inferior_median}{Threshold for lower limit of median ratio.}

\item{threshold_line_median}{Threshold line for median ratio.}

\item{n_samples}{Number of samples to label as outliers in plots.}

\item{save_csv}{Logical, whether to save the ratios as a CSV file. Default is FALSE.}

\item{csv_file}{The name of the CSV file to save the ratios if save_csv is TRUE. Default is "QC_results.csv".}

\item{show_heatmap}{Logical, whether to show the heatmap of potential outliers. Default is FALSE.}

\item{countsdata}{A data frame containing the HTG count data. The data must include probes that start with "^NC-|^POS-|^GDNA-|^ERCC-" for the function to work correctly.}
}
\value{
This function generates multiple plots displaying various QC metrics, saves an Excel file with all the ratios, and optionally creates a heatmap highlighting potential outlier samples. Additionally, it identifies and returns the most probable outliers based on the QC analysis.
}
\description{
This function performs various quality control (QC) checks. The QC checks include:
- QC0: Percentage of positive values greater than 4%.
- QC1: Library size greater than 7e+06.
- QC2: Negative control threshold greater than 0.045.
- QC3: Genomic DNA threshold greater than 0.02.
- QC4: ERCC threshold greater than 0.025.

These thresholds are tailored for the HTG EdgeSeq transcriptomic panel but can be adjusted as needed.

In addition to these plots, this function generates a data frame that can be saved as a .csv file (a preview of it will be shown). This data frame includes:
- The sum of each probe for each sample (total genes, positive, negative, gdna, and ercc)
- The ratio for each sample
- The size of each sample
- A column called PCA_genes which indicates the samples furthest from the center (you can specify how many samples to highlight as the furthest from the center)

This function also includes an optional heatmap to highlight potential outlier samples, which will be saved in a vector.

The plots will be shown in the console and saved in the current working directory.
}
\examples{
# Run the function with example data
HTG_QC(counts_data, save_csv = TRUE, show_heatmap = TRUE)

}
