% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03.3- HTG_quant_to_qual.R
\name{HTG_quant_to_qual}
\alias{HTG_quant_to_qual}
\title{HTG_quant_to_qual}
\usage{
HTG_quant_to_qual(
  data,
  column,
  threshold,
  above_label = "yes",
  below_label = "no"
)
}
\arguments{
\item{data}{A data frame containing the quantitative column to be transformed.}

\item{column}{The name of the quantitative column to be transformed.}

\item{threshold}{The threshold value to classify the quantitative values into qualitative categories.}

\item{above_label}{The label for values above the threshold.}

\item{below_label}{The label for values below or equal to the threshold.}
}
\value{
Returns the modified data frame with the specified column transformed into qualitative values.
}
\description{
This function converts quantitative columns in a data frame into qualitative ones based on a specified threshold.
}
\examples{
# Transform the "Ciclina" column in the "data" data frame with threshold 60
AnnotData <- HTG_quant_to_qual(AnnotData, "Ciclina", 60, "high", "low")

}
