% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTG_survival.R
\name{HTG_survival}
\alias{HTG_survival}
\title{Perform Survival Analysis}
\usage{
HTG_survival(
  variable_01,
  time,
  col_data,
  counts_data,
  DEA = NULL,
  res = NULL,
  method = "maxstat",
  genes_to_use = NULL,
  outliers = NULL,
  pattern = NULL,
  remove_outliers = TRUE,
  TME = NULL
)
}
\arguments{
\item{variable_01}{Character. The name of the survival event variable (e.g., "Recurrence_01").}

\item{time}{Character. The name of the time variable (e.g., "time_to_recurrence").}

\item{col_data}{Data frame. Annotation data containing the survival event and time variables.}

\item{counts_data}{Matrix. Raw count data for the genes.}

\item{DEA}{DESeqDataSet object or NULL. Pre-existing DESeqDataSet object. If NULL, a new DESeqDataSet will be created from `counts_data` and `col_data`. Which is necesary to obtain normalized_counts.}

\item{res}{Data frame or NULL. Result data with `padj` values to identify top genes. Used if `genes_to_use` is NULL.}

\item{method}{Method for determining cutoff (maxstat, median, or quartiles).}

\item{genes_to_use}{Character vector or NULL. Specific genes to use for survival analysis. Takes precedence over `res`.}

\item{outliers}{Character vector. Outlier identifiers to be removed from the analysis.}

\item{pattern}{Character. Pattern to match and remove certain rows from the count data (e.g., "^NC-|^POS-|^GDNA-|^ERCC-").}

\item{remove_outliers}{Logical. Indicates whether outliers should be removed. Defaults to TRUE.}

\item{TME}{Data frame or NULL. Data with gene expressions used if neither `res` nor `genes_to_use` are provided.}
}
\value{
Generates Kaplan-Meier survival plots and saves them as PDF files. The plots are saved in the current working directory.
}
\description{
This function performs survival analysis based on the provided data. The function uses parameters in a hierarchical manner:
if `genes_to_use` is provided, it uses those specific genes; if `res` is provided, it uses the top genes from `res`;
if `TME` is provided, it uses the genes from `TME`. If multiple parameters are provided, `genes_to_use` takes precedence over
`res`, and `res` takes precedence over `TME`.
}
\examples{

# USING DEA RESULTS:
HTG_survival(variable_01 = "Recurrence_01",
             time = "time_to_recurrence",
             col_data = AnnotData_tutorial,
             counts_data = counts_data_tutorial,
             res = res_tutorial,
             method = "maxstat",
             genes_to_use = NULL,
             TME = NULL,
             outliers = outliers_tutorial,
             pattern = "^NC-|^POS-|^GDNA-|^ERCC-",
             remove_outliers = TRUE)

# USING SPECIFIC GENES
HTG_survival(variable_01 = "Recurrence_01",
             time = "time_to_recurrence",
             col_data = AnnotData_tutorial,
             counts_data = counts_data_tutorial,
             res = NULL,
             method = "quartiles",
             genes_to_use = c("LCP1", "OMA1"),
             TME = NULL,
             outliers = outliers_tutorial,
             pattern = "^NC-|^POS-|^GDNA-|^ERCC-",
             remove_outliers = TRUE)

# USING TME DATA
HTG_survival(variable_01 = "Recurrence_01",
             time = "time_to_recurrence",
             col_data = AnnotData_tutorial,
             counts_data = counts_data_tutorial,
             res = NULL,
             method = "median",
             genes_to_use = NULL,
             TME = TME_data_tutorial$EPIC,
             outliers = outliers_tutorial,
             pattern = "^NC-|^POS-|^GDNA-|^ERCC-",
             remove_outliers = TRUE)

}
